requirements.txt
habitat-lab>=0.2.3
torch>=1.13.0
numpy>=1.21.0
gymnasium>=0.28.1
PyYAML>=6.0